fix: Загрузить автотесты, поменять структуру проекта
