+++ .github/workflows/statictest.yml
@@ -1,0 +1,35 @@
+name: go vet test
+
+on:
+  workflow_dispatch:
+  pull_request:
+  push:
+    branches:
+      - master
+      - main
+
+jobs:
+
+  statictest:
+    runs-on: ubuntu-latest
+    container: golang:1.17
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v2
+
+      - name: Download statictest binary
+        uses: robinraju/release-downloader@v1
+        with:
+          repository: Yandex-Practicum/go-autotests-bin
+          latest: true
+          fileName: statictest
+          out-file-path: .tools
+
+      - name: Setup autotest binary
+        run: |
+          chmod -R +x $GITHUB_WORKSPACE/.tools/statictest
+          mv $GITHUB_WORKSPACE/.tools/statictest /usr/local/bin/statictest
+
+      - name: Run statictest
+        run: |
+          go vet -vettool=$(which statictest) ./...
