+++ .gitignore
@@ -12,8 +12,8 @@
 *.out
 
 # Binaries
-cmd/gofermart/main
-cmd/gofermart/shortener
+cmd/gophermart/main
+cmd/gophermart/gophermart
 
 # Dependency directories (remove the comment below to include it)
 vendor/
