--- cmd/gofermart/README.md
@@ -1,3 +1,0 @@
-# cmd/gofermart
-
-В данной директории будет содержаться код, который скомпилируется в бинарное приложение
