+++ cmd/gophermart/README.md
@@ -1,0 +1,4 @@
+# cmd/gophermart
+
+В данной директории будет содержаться код накопительной системы лояльности, который скомпилируется в бинарное
+приложение.
